\name{Read.isi-package}
\alias{Read.isi-package}
\alias{Read.isi}
\docType{package}
\title{Access old data saved in fixed-width format based on ISI-formatted codebooks. For instance able to automatically read data from the World Fertility Survey (WFS).

}
\description{
Old statistical data was often stored in formats that are difficult to gain access to by modern statistical software. An example of this are the data-files of the `World Fertility Survey', which are stored in fixed-width format and accompanied by codebooks in a format developed by the International Statistical Institute. The read.isi package allows to gain access to these statistical data automatically, or to convert to codebook to SPSS-syntax.}
\details{
\tabular{ll}{
Package: \tab read.isi\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2008-06-27\cr
License: \tab Copyright (C) 2008 Rense Nieuwenhuis (email: contact@rensenieuwenhuis.nl).

This package, designed for use in R-Project, is free; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation; either version 2 of the License, or (at your option) any later version. One additional condition holds: all publications resulting from the use of the read.isi package must explicitly refer to this package.

This package is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.

You should have received a copy of the GNU General Public License along with this program; if not, write to the Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA \cr
}

}
\author{Rense Nieuwenhuis <contact@rensenieuwenhuis.nl>}

\references{More information on usage and background of the development can be found on www.rensenieuwenhuis.nl/r-project/my-functions/read-isi/}

\keyword{ package }
\keyword{ connection }

