\name{Read.isi-package}
\alias{Read.isi-package}
\alias{Read.isi}
\docType{package}
\title{
Automated access to old World Fertility Survey data saved in fixed-width format based on ISI-formatted codebooks.
}
\description{
Old statistical data was often stored in formats that are difficult to gain access to by modern statistical software. An example of this are the data-files of the `World Fertility Survey', which are stored in fixed-width format and accompanied by codebooks in a format developed by the International Statistical Institute. The read.isi package allows to gain access to these statistical data automatically, or to convert to codebook to SPSS-syntax.
}
\details{
\tabular{ll}{
Package: \tab Read.isi\cr
Type: \tab Package\cr
Version: \tab 0.5.1\cr
Date: \tab 2008-08-05\cr
License: \tab GPL-3\cr
}


}
\author{
Rense Nieuwenhuis

Maintainer: Rense Nieuwenhuis <contact@rensenieuwenhuis.nl>


}
\references{

For more information, manual, and background, please visit: www.rensenieuwenhuis.nl/r-project/read-isi
}


\keyword{ package }
\keyword{ connection }
\seealso{


}

